% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelcomp.R
\name{coef.comparison}
\alias{coef.comparison}
\title{Compare ergm coefficients}
\usage{
\method{coef}{comparison}(coeflist)
}
\description{
Takes in a list of multiple outputs from \code{\link{ergm.info}},
formatting it all into a table comparing up to 5 models.
}

