% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelcomp.R
\name{ergm.info}
\alias{ergm.info}
\title{ergm object information}
\usage{
ergm.info(object)
}
\description{
Takes an ergm object and gathers some of the information from
\code{\link{summary.ergm}}, in preparation to be passed to the function
\code{\link{coef.comparison}}.
}

