% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_sw.R
\name{run_sw}
\alias{run_sw}
\title{Run statnetWeb}
\usage{
run_sw()
}
\description{
Runs the statnetWeb shiny application, a GUI for the statnet
suite of network analysis packages.
}
\examples{
\dontrun{
run_sw()
}
}
\keyword{GUI,}
\keyword{shiny}

