% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statnetWeb-package.R
\docType{package}
\name{statnetWeb}
\alias{statnetWeb}
\alias{statnetWeb-package}
\title{statnetWeb: A graphical user interface for network analysis.}
\description{
The statnetWeb package makes it easy to run the graphical user interface
for the statnet suite of network analysis packages.
}

